Fix possible infinite recursion when retrieving the public callers of a method
Also, we longer perform an exhaustive search to find every possible
public caller, we rather cut short once the lengths of the visited paths
in the call graph exceed a certain threshold (Properties.MAX_RECURSION).

Furthermore, the computed results are now stored in a cache to allow for
more efficient retrieval.

Signed-off-by: Sebastian Schweikl <schweikl@posteo.de>