Fix NPE that occurred when the owner class of a public caller could not be reflected
The fix now tries all possible callers and chooses the first one that
works. To this end, the owner class of that caller must be reflected.
When a class cannot be reflected, the null reference is handed to us.
The fix now checks for null prior to working with the reflected class.

Signed-off-by: Sebastian Schweikl <schweikl@posteo.de>