Added new property FAVOR_SIMPLE_METHODS
If activated, the test generation process will try to cover simpler
methods (in terms of cyclomatic complexity) first. More specifically,
the goal to cover next is determined using a random-biased selection,
where the probability of a goal being selected is inversely proportional
to the cyclomatic complexity of the executable (i.e., method or ctor)
the goal is located in.

If deactivated, the goal to cover next is chosen randomly (every goal
has the same probability of being selected, regardless of its
complexity).

Signed-off-by: Sebastian Schweikl <schweikl@posteo.de>